Fix CMake builds from paths with strange characters
Fixes errors in CMake builds which occur when either the code and/or
build paths contain strange characters like whitespace, '&', ',', '(' and so on.

Please note that following characters may NOT be in the project/build
path anyway:
- '\', ';', '"', '\n' - causes CMake to error-out on it's own internal files;
- '<', '>', '#' - rejected by CMake in an error message;

For Windows this list will most likely also contain '/'.

Only ' ', '\t' and '\n' whitespace characters were tested (' ' and
'\t' work fine) - other invisible or non-printable characters may or may
not work.